Updates to buildDocs.sh
Use the name of the parent folder if the file is called index.js to produce a nice .md file name and have a better name in the heading of the markdown file.

for example:

src/components/Input/index.js

would return previously index.md and the heading of the md file to be Index.

new behaviour now reads Component.displayName = 'My Component'

returns MyComponent.md with heading being My Component

if displayName is not set we use the parent folder name if and only if the current file being processed is called index.js(x)

to allow for use case where

src/components/Input/index.js

returns Input.md and heading Input in md file.